stabilize autogen.sh
See GH #101

+++ autogen.sh
@@ -12,7 +12,8 @@
 
 set -x
 rm -rf autom4te.cache config.status
-if test -d .git; then
+if test -d .git -a -f build-aux/git-version-gen
+then
     git fetch --tags
     rm .tarball-version
     v=`build-aux/git-version-gen .tarball-version`
